backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0

# Config
pydantic-settings>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# DB + migrations
SQLAlchemy>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0
psycopg[binary]>=3.2.0,<4.0.0

# Auth (JWT + Password Hashing)
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.9,<1.0.0

# Worker (optional, for “resume-worthy” background jobs)
celery>=5.4.0,<6.0.0
redis>=5.0.0,<6.0.0

